requirements.txt
# FastAPI and async server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Data processing
pandas==2.1.4

# HTTP client for Ollama
httpx==0.27.0

# Environment variables
python-dotenv==1.0.0
